Plugin names and interface guards
This is broken at the moment, but it introduces plugin names to module
registration and a check that plugins can use to check that dependency
plugins are present.

It also starts to move interfaces from the upstream PluGeth into the
xplugeth package so that plugins can implement interface guards to
ensure they implement the interface in the same form that is expected
by PluGeth.