Add patchset build tag where necessary
The producer plugin relies on StateDB.GetTrie() being present.
It's available in Foundation Geth, and we patch it in for ETC and
Bor with patchsets, but we need to be able to compile get_patchset.go
for that to work. Compiling get_patchset.go fails because StateDB.GetTrie()
doesn't exist until after we've patched, which we can't do until we've
compiled get_patchset.go.

To solve this chicken or egg problem, we introduce a patchset build
tag, which we pass in only when building get_patchset.go. The producer
plugin has an alternate version of triedump.go that provides stubs
that don't rely on patched interfaces.

We'll need to establish (and document) this as a pattern for plugins
that rely on those kinds of extensions, that those extensions will
need to be implemented in files that aren't included with the patchset
flag.